[Android] Fix randomly fail issue when copying files for app template
Because in the gyp system, the "copy" action may be executed in parallel,
which breaks the assumption of target "prepare_xwalk_app_template".

So create a new target "prepare_xwalk_app_template_from_chromium", and make
"prepare_xwalk_app_template_from_xwalk" depend on this target.
That would help to serialize the "copy" sequence.

BUG=https://github.com/crosswalk-project/crosswalk/issues/734