XWalkExtensionClient: do not warn in expected situations
Once XWalkExtensionClient decides to destroy an instance, it deletes
it but keeps it's id around waiting for server confirmation. The id is
associated with a NULL instance in the map.

When we send the DestroyInstance message to the server, it's already
marked that way in the client. So we may receive a pending PostMessage
from the server. In this case, we shouldn't warn, but simply ignore,
since there's no JS to send the message anymore.

We keep the warning for the case where after server confirmed the
destruction, we get another message. This will indicate a real problem
either in extension implementation or in the system.