[Android] Keep the first extension instance id instead of the last one
It is a temporary policy for hanlding multiple extension instances on
Android. The potential problem of keeping the id of the last extension
instance created by native side is, in case of the second WebContents
is opened by the first WebContents (e.g. an async API calling), the reply
message for API calling might be incorrectly routed to the second one.