[Android] make_apk.py should exit before it calls others when failed.
Because the current process only get the output of subprocess, but can't
get the exception from the subprocess, so the parent process can't check
whether subprocess failed, and continue to run the following script.

This patch forces the parent process to read the returncode of subprocess
before continue execution, and direct exit when the returncode doesn't
equal to 0.

BUG=https://github.com/crosswalk-project/crosswalk/issues/789