[Android] Add the instrumentation test for runtime client.
This patch is the implementation of instrumentation test for
runtime client, it tests the Runtime APIs in the file "XWalkRuntimeClient.java".
This patch contains the test about embedded and shared client.

Implement a callback function in the test client, transfer it to
XWalkCoreProvider, then set it to the XWalkClient.
In the XWalkClient's callback, invoke the callback functions
which we defined, notify the test client about the page status.

For embedded client test:
    1. install XWalkRuntimeClientEmbeddedShell.apk
    2. python build/android/test_runner.py instrumentation \
       --test-apk=XWalkRuntimeClientEmbeddedTest -v

For shared client test:
    1. install XWalkRuntimeLib.apk
    2. install XWalkRuntimeClientShell.apk
    3. python build/android/test_runner.py instrumentation \
       --test-apk=XWalkRuntimeClientTest -v