[Extensions] Add --enable-loading-extensions-on-demand switch
Runtime switch to protect the implementation of on demand loading of
extensions ("lazy load") until it's done, when we change the semantics
to be --disable.

In the release this feature debuts, we keep the disable switch to allow
QA disable if needed. After we succesfully ship one stable Crosswalk
with the feature (it will be Crosswalk 2), the current plan is to remove
the switch, because we don't intend to keep the codepath for the
non-lazy behavior.