[Android] The first step to fix the resource id generation for core library.
Currently the resource ids for org.chromium.ui and org.chromium.content
are not generated so the web app will crash when running. Because
the container can't get the resources and their ids.
This patch is the first step to copy necessary resources from these two
targets to xwalk core library. The next step is to generate the correct
R.java.

BUG=#862