[Add] Add new APIs to XWalkClient for renderer responsiveness
This fix is to add 2 new APIs to XWalkClient for renderer
responsiveness. Internally, xwalk will start a timer to check if the
renderer become unresponsive if no ACK message for handling input event
is received from the renderer in 30 seconds. In this case, the callback
onRendererUnresponsive will be called. If it becomes responsive again,
the callback onRendererResponsive will be called.

BUG=https://github.com/crosswalk-project/crosswalk/issues/913
TEST=org.xwalk.core.xwview.test.RendererResponsivenessTest