[Extensions] Verify that the entry points are valid names
Also make sure that there aren't any extension already registered
with clashing names (or entry points).

We re-use the function that used to validate extension names to validate
entry points as well, so we rename it to something more appropriate.

All the symbols "exported" by an extension are kept on a std::set, so we
can re-use this Set when testing the extension name.