[Extensions] Remove GetJavaScriptAPI() virtual function
XWalkExtension expected subclasses to override a virtual function
GetJavaScriptAPI() that would return a "const char*" null terminated
string.

This patch removes that function and adds a set_javascript_api() for
subclasses to call during construction time. XWalkExtension stores the
javascript_api code in a similar way than it does for the name.

In practice, this matches what the external extension adapter was doing
anyway, and remove a few lines of boilterplate from every test case. It
will also be helpful when we start to use ResourceBundle for JS files
inside Crosswalk.