[Extensions] Load extensions on demand
Now that we have all the pieces it is time to put them together.

For each extension marked as 'should use the trampoline', we setup
"trap" acessors (for all entry points), that when run will load the
extension and create an instance for it. Before the extension is
loaded we need to remove all the references to the accessor function,
so it is not run anymore for those objects.