[Android] Add the instrumentation test for external extension.
This patch is to create two unit tests for external extensions for runtime
client. Since there are two modes - shared and embedded modes which are
supported in runtime client, here two copies here.

To avoid duplicated data, a shared directory is created to store data and
some code which should be used by all other instrumentation tests. So
we don't need to copy them everywhere. All testing data should be put
in here in future.

Besides, an external Java extension implementation is needed and
it has to be in test shell apks together with some assets and can't be in
instrumentation test apks because test shells can't access assets in
other assets.

For future, we should avoid duplicated code for embedded shell and shared
shell. This is a TODO.

BUG=#986