[Android] Add the interface of loadAssetFile and onTitleChanged.
Add the interfaces in embedded and shared mode.
loadAssetFile: It is needed when we put the test files(eg. html) under
the test assets.
onTitleChanged: Implement a callback function in the test client,
transfer it to XWalkCoreProvider, then set it to the XWalkWebChromeClient,
In the XWalkWebChromeClient,'s callback, invoke the callback functions
which we defined, notify the test client about the title changed.
Move most of the util codes to a utility folder for reuse.