[Android] Fix the crash when running some WebAPI components
Root cause: The crash of these WebAPI test cases are all caused
by the failure to access the i10n strings loaded from pak file.
In shared apk mode, the pak should be loaded from the Library apk
instead of App apk, for which we should use the Context of Library
apk for ResourceExtractor(under Content layer) to extract the pak
file. But after rebasing to V31, ResourceExtractor gets the Context
by mContext.getApplicationContext, which results that it gets the
wrong Context of App apk.

Fix: Force to initilize Context of ResourceExtractor before the
browser process starts.

BUG=https://github.com/crosswalk-project/crosswalk/issues/941