[Android] Disable flaky tests in xwalk_core_test
The responsiveness tests seem to be failed randomly on buildbot. It might be
caused by the unstable status on testing phone since this test relies on a
timeout to wait for unresponsiveness in 40 seconds. However, there is no way
to ensure it would happen in 40 seconds so that the test can be passed. Locally,
there 2 tests can be passed.