[Extensions] Pass more data in EPH construction time
After changes to support multiple Extension Processes, we are creating
ExtensionProcessHost (EPH) right when RenderProcessHost is created, so
we can pass RPH directly to EPH. This simplify the logic of sending the
channel handle back.

We also pass the register extension path and send it right after the
process is launched, instead of creating a new task for this.

The members ep_rp_channel_handle_ and
is_extension_process_channel_ready_ are currently unnecessary but will
be helpful in follow up changes, so they remained.