[Android] Refactor the code for instrumentation tests.
Currently so many source codes are duplicated and hard to maintain.
Need to refactor them and share common code. This fix is to share
all common code in utils so that instrumentation tests for embedded
and shared modes can share as many code as possible.
This is the second step. The next one is to share the code for core
tests.

BUG=https://crosswalk-project.org/jira/browse/XWALK-95