[Extensions] Check for conflicting entry points in the ModuleSystem
As we may have extensions coming from various processes, think that internal
extensions are run in the Browser Process and external extensions are run in
the Extension Process, the only point when we have all information of all the
extensions that would be run for a context is the module system.

So, if conflicting entry points are found, we log this as an warning and do
not register the extension.