[Android] Add the test case about enableRemoteDebugging.
This patch contains the test case about enableRemoteDebugging on
embedded and shared mode. It's for the runtime client.
Enable the enableRemoteDebugging by the broadcast. From the
output of 'cat /proc/net/unix' command, find the abstract socket
name opened for remote debugging.

For embedded client test:
1. install XWalkRuntimeClientEmbeddedShell.apk
2. python build/android/test_runner.py instrumentation \
--test-apk=XWalkRuntimeClientEmbeddedTest -v -f testEnableRemoteDebugging

For shared client test:
1. install XWalkRuntimeLib.apk
2. install XWalkRuntimeClientShell.apk
3. python build/android/test_runner.py instrumentation \
--test-apk=XWalkRuntimeClientTest -v -f testEnableRemoteDebugging