[UI][Tizen] Create NativeAppWindowTizen
Instead of adding custom behavior into NativeAppWindowViews behind
ifdefs, we create a proper subclass for Tizen. This allow us to
implement observers directly in our window, for example the sensor
observer.

This is a preparation for allowing the window to tell the indicator
which orientation it should paint. In the previous design (using
TizenSensorObserver) we would end up adding more couple (or more layer
of indirection) to accomplish this.

There are some "ifdef" leftovers in NativeAppWindowViews, but can be
moved in later patches.