[Extensions] Route messages to the appropriated server
Based on the extension name and the extension ID, we route the messages
to the appropriated ExtensionServer running on the Extension thread or the
UI thread.

On the in-process extensions, 2 messages that are normally directed to the
ExtensionServer on the Extension Process are now completely handled by the
filter. One example is the GetExtensions message that has to combine a list
of available extensions from two different servers in a single reply.