[Tizen] Add support for incremental build
Current GBS build takes over 2 hours, this is not acceptable for developer
especially for trybot. This commit changed the spec file to skip its source code
clean and prepare steps, instead build directly in src tree on hosted system.
Thus the build is greatly speeded up.

The side effect is that the generated source rpm package is not intact. Since
incremental build is guarded by a parameter, the official buildbot works as
before(a full clean rebuild).