[Shared Runtime Process Model] Daemon implemented using D-Bus and a launcher is included also.
There are several functions exported in server side like launch,
install unsinstall and so on. Most of these exported functions are
just wrappers of application service's members. Which functions
should be exported is only dependent on the need of launcher.  In
Linux and Tizen, the launcher is acting as a foreground dummy
process to representative a web process living in the daemon, more
precisely the renderer/extension process bundles.  Xwalk's D-Bus
interface is internally used between daemon and launcher and thus is
not introspectable.