[Runtime] Add application test API and adjust API browser test
An internal extension for application API test is created to explicitly return
the API test result. For example, with this API we don't to need wait and test
the page title or Runtime count anymore, a call to xwalk.app.test.notifyPass()
or xwalk.app.test.notifyFail() will return the test result to runtime process.

The application API browser test will register this extension in setup phase.
And block wait for the test complete.