[Extensions] Remove service parameter from register test callback
XWalkExtensionService allows us to register a function to be called
every time a new XWalkExtensionServer is created, so tests can register
specific extensions. This function may be called multiple times because
we create a (in process) server per Render Process created.

The callback also received a service parameter, to be used for setting
the path for external extensions to be loaded. But the previous commit
changed how tests handle this, so the parameter can be removed.