[Extensions] Check if reply is valid after a SendSyncMessageToNative
When we try to send a sync message to an extension instance that
became invalid (i.e.: the extension process died) then we might
end up with a NULL reply. We should then check before converting
to a V8Value, thus avoiding the RenderProcess to crash.

That is the same behavior PostMessageCallback has nowadays.