[Application] Move handling of management commands to ApplicationSystem
This patch is more about moving this logic out of XWalkBrowserMainParts
to a more reasonable location. We may decide to move this handling code
later (to hid it more) when implementing the IPC-based handling of the
mode.

Another outcome is that the logic for launching applications (and all
its cases) is clearer now.