[Tizen] Permit the inclusion of Tizen Web UI Framework
This patch creates a Tizen-specific XWalkContentRendererClient that
implements the WillSendRequest hook.  This hook, as the name implies, is
called before every request is made from a particular WebKit::WebFrame.  By
tapping into this hook, it is possible to rewrite requests so that they have
the full file path.

Some measures are taken so that relative paths outside where the Tizen Web
UI Framework can't be referenced; also, only known extensions are permitted.