[Runtime] Fix RuntimeContext ownership and unify its creation
On Android RuntimeContext used to be a singleton which is wrong as content::ContentBrowserClient lifetime should be equal to the duration of the browser session, meaning its creation&deletion should be controlled by BrowserMainLoop.

This patch unifies the creation of RuntimeContext between android and others removing ugly '#ifdefs' and unneeded difference in the code path.