[DBus] Add PropertyExporter and DBusManager classes
This commit bootstraps the xwalk/dbus directory with two classes we
intend to use when implementing the '--run-as-service' mode in Linux.

The 'DBusManager' is a convenience class that packages the creation of
the DBus thread and the bus connection. There need to be only one of
those for the whole Crosswalk, so it should be passed around for the
subsystems that need it.

The 'PropertyExporter' class is a partial implementation of a helper
class that exports the 'org.freedesktop.DBus.Properties' interface for a
certain 'dbus::ExportedObject'. It's interface is still incomplete, and
at the moment it will be focused on covering our use cases for
'--run-as-service'. Chromium have a infrastructure for dealing with
properties but only on the client-side, this class adds the server side.

Unit testing infrastructure, as well as basic sanity tests were
added. With these bootstrapped is easy to add new tests later. The new
target 'xwalk_all_tests' was added to include this module in the build
system.