[Extension] Catch when RenderProcess dies.
This patch enables us to be warned when a RenderProcess dies through
Runtime::RenderProcessGone(), which is our WebContentObserver interface.

With that in place, we can then warn XWalkExtensionService so it can gracefully
shutdown the associated ExtensionProcess and in_process ExtensionServer. There
is no need to handle MessageFilters at this point anymore since they are owned
by the IPC channel. This, in turn, is deleted when the RenderProcess dies.