[Application] Add --run-as-service and ApplicationServiceProvider
When the switch '--run-as-service' is used, Crosswalk doesn't launch any
application, but only sits on its mainloop. The idea is that external
requests will happen to trigger app launches.

ApplicationServiceProvider will be the entity responsible for exposing
some interface for external requests. In Linux the current plan is to
expose a D-Bus service.

The Provider receives a pointer to ApplicationService, the idea behind
this design is that all the actual functionality should be implement in
ApplicationService, and the Provider will mostly bridge that
functionality through an external interface.