[Application] Add ApplicationServiceProvider for Linux
The ApplicationServiceProviderLinux will use a D-Bus service to expose
objects and let external environment manage applications as well as
launch applications.

For now a placeholder object that answers the message "Ping" with a
"Pong" response is there. Next commits will replace it with a real
object that exposes functionality from ApplicationService.