[Runtime] Make Runtimes know their application ID
Crosswalk is switching its process model to shared runtime process mode, meaning that we are going to have several applications sharing the same browser process.

This patch makes a runtime know which application it belongs to. This data further will be used in Runtime registry and other classes as they have to be modified in order to support shared runtime process mode.