[Android] Add the test cases about onActivityResult.
This patch contains the test case about onActivityResult in shared
and embedded mode. It's for the runtime client.
Create a activity names "SecondActivity". In the runtime client activity,
it starts activity for result. In the second activity, it will create a new
intent, and transfer data to runtime client activity.