Added permission group and policy manageemnt classes.
Added two major components of crosswalk API permission:
The permission group management and the permission policy management.
The permission group management is responsible for the mapping between
permission name and corresponding function name. The permission policy
manager is used to calculate the permission for a request.
For more details please refer to the design doc:
https://docs.google.com/a/intel.com/document/d/1TfU_oZo6P2Ff24w5RjRhYfPTJae5EzRWiXtdtxz0yBo/edit#